Move WorkEffortsFilterPeriod enum to filter.types
