+++ alv-portal-ui/src/app/online-services/work-efforts/work-efforts/work-efforts-filter.types.ts
@@ -1,7 +1,14 @@
 import {
-  WorkEffortsFilterPeriod,
   WorkEffortApplyStatus
 } from '../../../shared/backend-services/work-efforts/proof-of-work-efforts.types';
+
+export enum WorkEffortsFilterPeriod {
+  CURRENT_MONTH = 'CURRENT_MONTH',
+  LAST_3_MONTHS = 'LAST_3_MONTHS',
+  LAST_6_MONTHS = 'LAST_6_MONTHS',
+  LAST_12_MONTHS = 'LAST_12_MONTHS',
+  ALL_MONTHS = 'ALL_MONTHS'
+}
 
 export interface WorkEffortsFilterValues {
   period: WorkEffortsFilterPeriod;
